# Paralellise the shmup converter