--- a/Cargo.toml
+++ b/Cargo.toml
@@ -37,6 +37,7 @@ bevy-inspector-egui = "0.15.0"
 rand = "0.8.5"
 sha2 = "0.10.6"
 data-encoding = "2.3.3"
+rayon = "1.6.1"
 
 [dev-dependencies]
 assert_cmd = "2.0"