Here is a benchmark on the repository https://github.com/kubernetes/kubernetes

The test was run on a M2 Macbook Pro 

```
 Chip: Apple M1 Max
 Total Number of Cores:	10 (8 performance and 2 efficiency)
 Memory: 64 GB
```

Before ~= 1:52:79
```console
$time cargo run --features bevy/dynamic scan https://github.com/kubernetes/kubernetes
cargo run --features bevy/dynamic scan   110.53s user 2.28s system 100% cpu 1:52.79 total
```

After ~= 17.140
```console
$time cargo run --features bevy/dynamic scan https://github.com/kubernetes/kubernetes
cargo run --features bevy/dynamic scan   142.36s user 3.86s system 853% cpu 17.140 total
```